Chan ge the GlobalConfig object to a Singleton pattern, changes here reflect using the singleton.
Add pushover emergency notification support (PoC) - NEEDS REFINING - specifically the timerange.

Depending on the timerange and when the script executes will vary the meaning of "yesterday at 8 PM" greatly. Need to figure out a way to make that statement work without having to resort to a formatted date or editing config file daily!