Wording, small fixes.
There seems to be an error when trying to create tokens.txt on first start-up. install.sh has been modified to copy tokens.txt empty file into the ./push directory. No idea why, error is 'permission denied' when the user running the Perl daemon is www-data and the directories are owned bvy www-data.